refactor: export createFetchival
